评论列表

+++ JBlog-JPA/src/main/java/com/youthlin/jblog/model/Page.java
@@ -9,15 +9,15 @@ import java.util.List;
  * 分页
  */
 public class Page<T> implements Serializable {
-    private long pageIndex = 0;//当前第几页
-    private long pageTotal = 1L;//共几页
-    private long countPerPage = 5;//每页多少个
+    private int pageIndex = 0;//当前第几页
+    private int pageTotal = 1;//共几页
+    private int countPerPage = 5;//每页多少个
     private List<T> item = new ArrayList<>(0);
 
     public Page() {
     }
 
-    public Page(long pageIndex, long pageTotal, long countPerPage, List<T> item) {
+    public Page(int pageIndex, int pageTotal, int countPerPage, List<T> item) {
         this.pageIndex = pageIndex;
         this.pageTotal = pageTotal;
         this.countPerPage = countPerPage;
@@ -34,27 +34,27 @@ public class Page<T> implements Serializable {
                 '}';
     }
 
-    public long getPageIndex() {
+    public int getPageIndex() {
         return pageIndex;
     }
 
-    public void setPageIndex(long pageIndex) {
+    public void setPageIndex(int pageIndex) {
         this.pageIndex = pageIndex;
     }
 
-    public long getPageTotal() {
+    public int getPageTotal() {
         return pageTotal;
     }
 
-    public void setPageTotal(long pageTotal) {
+    public void setPageTotal(int pageTotal) {
         this.pageTotal = pageTotal;
     }
 
-    public long getCountPerPage() {
+    public int getCountPerPage() {
         return countPerPage;
     }
 
-    public void setCountPerPage(long countPerPage) {
+    public void setCountPerPage(int countPerPage) {
         this.countPerPage = countPerPage;
     }
